first stable version

+++ ch2_snake_model/include/gradient_descent_base.h
@@ -11,13 +11,13 @@ class GradientDescentBase {
     virtual void update() = 0;
     virtual bool is_terminate(int current_iter, int max_iteration) const;
 
-    virtual double compute_energy() = 0;
+    virtual float compute_energy() = 0;
     virtual void roll_back_state() = 0;
     virtual void back_up_state() = 0;
     virtual void update_step_size(bool is_energy_decent);
 
     virtual void print_terminate_info() const;
 
-    double step_size_ = 1e-10;
-    double last_energy_ = std::numeric_limits<double>::max();
+    float step_size_ = 1e-10;
+    float last_energy_ = std::numeric_limits<float>::max();
 };
